+++ NewEngine.h
@@ -8,6 +8,6 @@ void ProcessMessage();
 void SetWindowSize(int WIN_WIDTH, int WIN_HEIGHT);
 void SetWindowTitle(const wchar_t* TITLE);
 bool CloseWindow();
-void SetBackgroundColor(int Red, int Green, int Blue);
+void SetBackgroundColor(float Red, float Green, float Blue);
 void NewEnginePreDraw();
 void NewEnginePostDraw();
